+++ Polly.Contrib.AzureFunctions/CircuitBreaker/DurableCircuitBreakerEntity.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Threading.Tasks;
-using System.Xml;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
 
@@ -150,45 +149,10 @@ namespace Polly.Contrib.AzureFunctions.CircuitBreaker
             // The first time the circuit-breaker is accessed, it will self-configure.
             if (breakerState == null)
             {
-                breakerState = ConfigureCircuitBreaker(context, log);
+                breakerState = ConfigurationHelper.ConfigureCircuitBreaker(context, log);
             }
 
             return breakerState;
         }
-
-        private static BreakerState ConfigureCircuitBreaker(IDurableEntityContext context, ILogger log)
-        {
-            string circuitBreakerId = context.Self.EntityKey;
-            log.LogCircuitBreakerMessage(circuitBreakerId, $"Setting configuration for circuit-breaker {circuitBreakerId}.");
-
-            // Intentionally no defaults - users should consciously decide what tolerances suit the operations invoked through the circuit-breaker.
-            TimeSpan breakDuration = XmlConvert.ToTimeSpan(GetCircuitConfiguration(circuitBreakerId, "BreakDuration"));
-            int maxConsecutiveFailures = Convert.ToInt32(GetCircuitConfiguration(circuitBreakerId, "MaxConsecutiveFailures"));
-
-            var breakerState = new BreakerState
-            {
-                CircuitState = CircuitState.Closed,
-                BrokenUntil = DateTime.MinValue,
-                ConsecutiveFailureCount = 0,
-                MaxConsecutiveFailures = maxConsecutiveFailures,
-                BreakDuration = breakDuration
-            };
-
-            if (breakerState.BreakDuration <= TimeSpan.Zero)
-            {
-                throw new InvalidOperationException("The circuit-breaker must be configured with a positive break-duration.");
-            }
-
-            if (breakerState.MaxConsecutiveFailures <= 0)
-            {
-                throw new InvalidOperationException("The circuit-breaker must be configured with a max number of consecutive failures greater than or equal to 1.");
-            }
-
-            context.SetState(breakerState);
-            return breakerState;
-        }
-
-        internal static string GetCircuitConfiguration(string circuitBreakerId, string configurationItem)
-            => Environment.GetEnvironmentVariable($"CircuitBreakerSettings:{circuitBreakerId}:{configurationItem}", EnvironmentVariableTarget.Process);
     }
 }